moved strategy factory to other package
