+++ src/test/java/lspMobsimTests/MultipleShipmentsCompleteLSPMobsimTest.java
@@ -20,7 +20,7 @@
 
 package lspMobsimTests;
 
-import example.lsp.multipleChains.AssignmentStrategyFactory;
+import example.lsp.lspReplanning.AssignmentStrategyFactory;
 import lsp.*;
 import lsp.resourceImplementations.distributionCarrier.DistributionCarrierUtils;
 import lsp.resourceImplementations.collectionCarrier.CollectionCarrierUtils;
